Extract class for rendering domain templates
So can do this outside of commands for configuring specific Metalware
items.
